Improve mod asset loading
Previously if two mods had an asset with the same name they would fail
to load correctly because the asset path prefix was ignored. The asset
path prefix is now respected and both assets are loaded correctly.

Additionally the asset bundle manifest is now used to resolve the path
to the asset to avoid loading files that are not referenced from the
manifest.

The asset prefab name is now used to improve tool specification
override matching.